remove MegaMoE module and merge bench_dispatch_grouped_gemm.py bench_grouped_gemm_combine.py into one
