+++ benchmark/ops/training/bench_mega_moe.py
@@ -4,34 +4,54 @@
 # See LICENSE for license information.
 ###############################################################################
 
-"""Shared helpers for the mega-MoE EP benchmarks (dispatch+GEMM / GEMM+combine).
+"""Benchmark the fused BF16 mega MoE kernels (EP, intra-node).
 
-Single source of truth for: routing/weight generation, the dense-GEMM roofline,
-the CUDA-event bench helper, the prologue-driven input builders for both kernels,
-and the per-stage metric/print template so both benchmarks print identically.
+Two modes, selected with --mode:
+  dispatch_grouped_gemm : fused dispatch + grouped GEMM
+  grouped_gemm_combine  : fused grouped GEMM + combine
+
+Self-contained: the shared reference ops (routing/weight generation, the
+dense-GEMM roofline, the CUDA-event bench helper, the prologue-driven input
+builders, and the per-stage metric/print template) live in this file too.
 """
 
+import argparse
 import functools
 import math
 import os
 import statistics
+import sys
+from dataclasses import dataclass, field
+from datetime import datetime
 from types import SimpleNamespace
-from typing import NamedTuple
+from typing import Any, Callable, NamedTuple
 
-import flydsl.compiler as flyc
-import flydsl.expr as fx
 import numpy as np
+import pandas as pd
 import torch
 import torch.distributed as dist
-from flydsl.expr import arith
-from flydsl.expr.buffer_ops import (
+from tabulate import tabulate
+
+# config + this module's helpers are same-dir (auto on the script-dir path)
+from config import gen_moe_test_cases, get_platform_info
+
+# repo root (primus_turbo) on the path
+_HERE = os.path.dirname(__file__)
+sys.path.insert(0, os.path.abspath(os.path.join(_HERE, "..", "..", "..")))
+
+import flydsl.compiler as flyc  # noqa: E402
+import flydsl.expr as fx  # noqa: E402
+from flydsl.expr import arith  # noqa: E402
+from flydsl.expr.buffer_ops import (  # noqa: E402
     buffer_load,
     create_buffer_resource,
     extract_base_index,
 )
-from flydsl.expr.typing import AddressSpace, PointerType
+from flydsl.expr.typing import AddressSpace, PointerType  # noqa: E402
 
-from primus_turbo.flydsl.gemm.gemm_bf16_kernel import (
+# primus_turbo.flydsl.* imported before primus_turbo.pytorch (kept from the
+# original mega_utils order; the two fused kernels below need pytorch first).
+from primus_turbo.flydsl.gemm.gemm_bf16_kernel import (  # noqa: E402
     _compile_dense_nt,
     _compile_grouped_variable_k_bf16,
     _get_compiled_dense,
@@ -39,11 +59,11 @@ from primus_turbo.flydsl.gemm.gemm_bf16_kernel import (
     _make_shared_storage,
     gemm_bf16_tile,
 )
-from primus_turbo.flydsl.mega import (
+from primus_turbo.flydsl.mega import (  # noqa: E402
     dispatch_prologue_flydsl_kernel,
     swiglu_flydsl_kernel,
 )
-from primus_turbo.flydsl.mega.ep_intranode import (
+from primus_turbo.flydsl.mega.ep_intranode import (  # noqa: E402
     _BLOCK_THREADS,
     _NUM_WARPS,
     _PVEC,
@@ -51,52 +71,28 @@ from primus_turbo.flydsl.mega.ep_intranode import (
     dispatch_bf16_tile,
     topk_reduce_bf16_tile,
 )
-from primus_turbo.flydsl.mega.symm_buffer import (
+from primus_turbo.flydsl.mega.symm_buffer import (  # noqa: E402
     BLOCK_M as _POOL_BLOCK_M,
 )
-from primus_turbo.flydsl.mega.symm_buffer import (
+from primus_turbo.flydsl.mega.symm_buffer import (  # noqa: E402
     TOKEN_DTYPE,
     SymBuffer,
     Workspace,
     get_symm_buffer_for_mega_moe,
 )
-from primus_turbo.flydsl.utils.gemm_helper import (
+from primus_turbo.flydsl.utils.gemm_helper import (  # noqa: E402
     ceildiv,
     make_value_attrs,
     xcd_remap_pid,
 )
 
-# re-export so benchmarks can import everything mega from one place
-__all__ = [
-    "generate_routing",
-    "dense_gemm_peak_ms",
-    "bench",
-    "gate3",
-    "check_accuracy",
-    "AccuracyCheck",
-    "generate_input",
-    "compute_stage_metrics",
-    "BF16_BYTES",
-    "sync_ranks",
-    "reduce_across_ranks",
-    "aggregate_stage_metrics",
-    "stage_columns",
-    "turbo_csv_row",
-    "checks_verdict",
-    "apply_case",
-    "all_ranks_ok",
-    "print_header",
-    "print_stage",
-    "dispatch_only",
-    "dispatch_prologue_flydsl_kernel",
-    "grouped_gemm_bf16_only",
-    "grouped_gemm_variable_k_only",
-    "compile_grouped_gemm_bf16",
-    "get_symm_buffer_for_mega_moe",
-    "swiglu_flydsl_kernel",
-    "combine_only",
-    "topk_reduce_only",
-]
+# import primus_turbo.pytorch first to dodge the mega kernels' circular import
+import primus_turbo.pytorch  # noqa: E402,F401
+from primus_turbo.flydsl.mega import (  # noqa: E402
+    dispatch_grouped_gemm_bf16_flydsl_kernel,
+    grouped_gemm_combine_bf16_flydsl_kernel,
+)
+from primus_turbo.pytorch.ops import grouped_gemm as turbo_grouped_gemm  # noqa: E402
 
 
 # --------------------------------------------------------------------------- #
@@ -1125,3 +1121,731 @@ def topk_reduce_only(
         stream=torch.cuda.current_stream(),
     )
     return output
+
+
+# Fast-skip flydsl tiling gaps: moe_intermediate_size not a multiple of GEMM tile (BN=256) -> autotune fails. TODO(mega): add tail N tile.
+UNSUPPORTED = {"DeepSeek-V2-Lite", "MoE-1T"}
+
+
+###############################################################################
+# Shared scaffolding (both modes)
+###############################################################################
+
+
+@dataclass
+class StageMetrics:
+    """Raw per-rank timings + work for one stage (fields match compute_stage_metrics kwargs)."""
+
+    gemm_ms: float
+    dense_ms: float
+    dense_gm: int
+    comm_ms: float
+    fused_ms: float
+    flops: float
+
+
+@dataclass
+class StageSpec:
+    """Declarative spec for one benchmarked stage; fused_fn is reused for timing + accuracy check."""
+
+    name: str
+    flops: float
+    dense_dims: tuple[int, int, int]
+    gemm_fn: Callable
+    comm_fn: Callable  # comm baseline (dispatch_only / combine_only)
+    fused_fn: Callable
+    ref_fn: Callable
+    acc_slice: slice = slice(None)
+
+
+class StageRunner:
+    """Binds per-run context so each stage passes only its own knobs; run does the shared template."""
+
+    def __init__(self, group, args, synced_fn, group_m_cands):
+        self.group = group
+        self.args = args
+        self.synced_fn = synced_fn
+        self.group_m_cands = group_m_cands
+
+    def run(self, spec):
+        """Time gemm / dense roofline / comm_only / fused, then gate accuracy under a synced bracket."""
+        args = self.args
+        dense_m, dense_n, dense_k = spec.dense_dims
+        t_gemm = bench(spec.gemm_fn, iters=args.iters)
+        t_dense, dense_gm = dense_gemm_peak_ms(
+            dense_m, dense_n, dense_k, 256, 256, args.iters, group_m_cands=self.group_m_cands
+        )
+        t_comm = bench(spec.comm_fn, iters=args.iters)
+        t_fused = bench(spec.fused_fn, iters=args.iters)
+        metrics = StageMetrics(
+            gemm_ms=t_gemm,
+            dense_ms=t_dense,
+            dense_gm=dense_gm,
+            comm_ms=t_comm,
+            fused_ms=t_fused,
+            flops=spec.flops,
+        )
+        # accuracy: fused vs ref over the SAME symm state (both under a synced bracket)
+        ref_out = self.synced_fn(spec.ref_fn)
+        out = self.synced_fn(spec.fused_fn)[0]
+        check = check_accuracy(self.group, spec.name, out[spec.acc_slice], ref_out[spec.acc_slice])
+        return metrics, check
+
+
+def _make_runner(group, args):
+    """Build a StageRunner with a synced-bracket probe wrapper (shared by both modes)."""
+
+    # synced bracket isolates accuracy probes from timing; scoreboard is never-reset parity
+    def _synced(run_fn):
+        sync_ranks(group)
+        out = run_fn()
+        sync_ranks(group)
+        return out
+
+    group_m_cands = (args.dense_group_m,)
+    return StageRunner(group, args, _synced, group_m_cands)
+
+
+###############################################################################
+# Mode: dispatch_grouped_gemm
+###############################################################################
+
+
+@dataclass
+class DispatchContext:
+    """Shared per-run state built by _dispatch_make_context: config, input/symm buffers, derived geometry."""
+
+    # collective + CLI config
+    group: Any
+    args: Any
+    rank: int
+    # dispatched input namespace + symmetric buffers (own x / handle / pool / cap)
+    inp: Any
+    symm: Any
+    # derived geometry: expert-major pool, each expert padded to a BM multiple
+    M_eff: int  # total padded pool rows this rank GEMMs
+    experts_per_rank: int
+    padded_group_lens: Any  # per-expert BM-padded row counts, sum == M_eff
+    group_offs: Any  # cumulative padded boundaries [experts_per_rank + 1]
+
+
+def _dispatch_make_context(group, args):
+    """Build input + symm buffers and precompute shared geometry (weights built in generate_input)."""
+    inp = generate_input(
+        group,
+        kind="dispatch",
+        T=args.num_tokens,
+        H=args.hidden,
+        I=args.inter,
+        E=args.num_experts,
+        K=args.num_topk,
+        BLOCK_M=256,
+        BLOCK_N=256,
+    )
+    experts_per_rank = args.num_experts // group.size()
+    # BM-padded expert-major per-expert row counts (padding rows zero) so turbo gg matches the layout
+    real_tiles = int(inp.num_tile_blocks[0].item())
+    M_eff = real_tiles * 256
+    tile_experts = inp.tile_to_expert[:real_tiles].to(torch.int64)
+    counts = torch.bincount(tile_experts, minlength=experts_per_rank)[:experts_per_rank]
+    padded_group_lens = (counts * 256).to(torch.int64)  # sum == M_eff
+    # block_m-padded per-expert boundaries for the variable-K wgrads
+    group_offs = torch.zeros(experts_per_rank + 1, dtype=torch.int32, device="cuda")
+    group_offs[1:] = padded_group_lens.to(torch.int32).cumsum(0)
+    return DispatchContext(
+        group=group,
+        args=args,
+        rank=group.rank(),
+        inp=inp,
+        symm=inp.symm,
+        M_eff=M_eff,
+        experts_per_rank=experts_per_rank,
+        padded_group_lens=padded_group_lens,
+        group_offs=group_offs,
+    )
+
+
+def _dispatch_make_fused_call(ctx, lhs, rhs, layout, *, trans_c=False):
+    """Build the fused dispatch+GEMM call; stages differ only in operands / layout / trans_c."""
+    return lambda: dispatch_grouped_gemm_bf16_flydsl_kernel(
+        lhs,
+        rhs,
+        ctx.group,
+        handle=ctx.inp.handle,
+        layout=layout,
+        BM=256,
+        BN=256,
+        trans_c=trans_c,
+    )
+
+
+def _dispatch_make_comm_call(ctx, operand):
+    """Build the dispatch-only call (the comm baseline for one stage)."""
+    return lambda: dispatch_only(operand, ctx.inp.handle, ctx.symm)
+
+
+def _dispatch_stage_fwd(runner, ctx):
+    """forward (NT): N=2I, K=H; returns (metrics, check, xgmi_bytes)."""
+    inp, args = ctx.inp, ctx.args
+    pool = ctx.symm.dispatch_token_pool
+    M_eff, N_fwd, K = ctx.M_eff, 2 * args.inter, args.hidden
+    flops = 2.0 * M_eff * N_fwd * K
+    # XGMI push bytes per rank = remote rows (dest != rank) x hidden x bf16
+    dest_cpu, count_cpu = inp.destination.cpu(), inp.count.cpu()
+    remote_rows = int(count_cpu[dest_cpu != ctx.rank].sum().item())
+    xgmi_bytes = remote_rows * args.hidden * BF16_BYTES
+
+    spec = StageSpec(
+        name="fwd fused (nt)",
+        flops=flops,
+        dense_dims=(M_eff, N_fwd, K),
+        gemm_fn=lambda: grouped_gemm_bf16_only(
+            pool,
+            inp.W1,
+            inp.l1_out,
+            inp.tile_to_expert,
+            inp.num_tile_blocks,
+            BLOCK_M=256,
+            BLOCK_N=256,
+        ),
+        comm_fn=_dispatch_make_comm_call(ctx, inp.x),
+        fused_fn=_dispatch_make_fused_call(ctx, inp.x, inp.W1, "nt"),
+        ref_fn=lambda: turbo_grouped_gemm(
+            pool[:M_eff].contiguous(), inp.W1, ctx.padded_group_lens, trans_b=True
+        ),
+        acc_slice=slice(0, M_eff),
+    )
+    metrics, check = runner.run(spec)
+    return metrics, check, xgmi_bytes
+
+
+def _dispatch_stage_bwd_dgrad(runner, ctx):
+    """backward dgrad (NN): dispatch dy + L2 dgrad pool[M,H] @ w2 -> d_swiglu[M,I]; N=I, K=H."""
+    inp, args, symm = ctx.inp, ctx.args, ctx.symm
+    pool = symm.dispatch_token_pool
+    M_eff, N_bwd, K = ctx.M_eff, args.inter, args.hidden
+    flops = 2.0 * M_eff * N_bwd * K
+    dy = torch.ones(args.num_tokens, args.hidden, device="cuda", dtype=torch.bfloat16)
+    d_swiglu = torch.empty(symm.num_max_pool_tokens, N_bwd, device="cuda", dtype=torch.bfloat16)
+    # fill the pool with dy once so the gemm-only baseline contracts the right rows
+    sync_ranks(ctx.group)
+    dispatch_only(dy, inp.handle, symm)
+    sync_ranks(ctx.group)
+
+    # accuracy: fused dispatch+GEMM(NN) vs turbo grouped_gemm. NN -> trans_b=False.
+    spec = StageSpec(
+        name="bwd dgrad fused (nn)",
+        flops=flops,
+        dense_dims=(M_eff, N_bwd, K),
+        gemm_fn=lambda: grouped_gemm_bf16_only(
+            pool,
+            inp.W2,
+            d_swiglu,
+            inp.tile_to_expert,
+            inp.num_tile_blocks,
+            layout="nn",
+            BLOCK_M=256,
+            BLOCK_N=256,
+        ),
+        comm_fn=_dispatch_make_comm_call(ctx, dy),
+        fused_fn=_dispatch_make_fused_call(ctx, dy, inp.W2, "nn"),
+        ref_fn=lambda: turbo_grouped_gemm(
+            pool[:M_eff].contiguous(), inp.W2, ctx.padded_group_lens, trans_b=False
+        ),
+        acc_slice=slice(0, M_eff),
+    )
+    return runner.run(spec)
+
+
+def _dispatch_stage_bwd_wgrad(runner, ctx):
+    """backward wgrad dW1 (TN, variable-K): dW1 = pool(x)^T @ grad_l1; OUT_M=H, OUT_N=2I."""
+    inp, args, symm = ctx.inp, ctx.args, ctx.symm
+    pool = symm.dispatch_token_pool
+    cap = symm.num_max_pool_tokens
+    M_out, N_out = args.hidden, 2 * args.inter  # dW1: lhs feature = H (pool x), rhs feature = 2I (grad_l1)
+    flops = 2.0 * ctx.M_eff * M_out * N_out
+    x_pool = torch.ones(cap, M_out, device="cuda", dtype=torch.bfloat16)
+    grad_pool = torch.ones(cap, N_out, device="cuda", dtype=torch.bfloat16)
+    # trans_c: dW1 stored transposed [G, N_out, M_out] = [G, 2I, H] = W1-native layout
+    dW1 = torch.empty(ctx.experts_per_rank, N_out, M_out, device="cuda", dtype=torch.bfloat16)
+
+    ref_dW1 = torch.zeros_like(dW1)  # empty groups -> 0 (matches the fused padded output)
+    offs_cpu = ctx.group_offs.tolist()
+
+    def _ref():
+        dispatch_only(x_pool, inp.handle, symm)
+        sync_ranks(ctx.group)
+        for expert in range(ctx.experts_per_rank):
+            start, end = offs_cpu[expert], offs_cpu[expert + 1]
+            if end > start:  # padded rows are zero -> contract to 0 (skip is equivalent)
+                ref_dW1[expert] = (grad_pool[start:end].float().T @ pool[start:end].float()).to(dW1.dtype)
+        return ref_dW1
+
+    spec = StageSpec(
+        name="wgrad dW1 fused (tn)",
+        flops=flops,
+        # dense roofline of the same total FLOPs (one [H,2I] GEMM contracting M_eff rows)
+        dense_dims=(M_out, N_out, ctx.M_eff),
+        gemm_fn=lambda: grouped_gemm_variable_k_only(
+            x_pool, grad_pool, ctx.group_offs, dW1, BLOCK_M=256, BLOCK_N=256, trans_c=True
+        ),
+        comm_fn=_dispatch_make_comm_call(ctx, x_pool),
+        fused_fn=_dispatch_make_fused_call(ctx, x_pool, grad_pool, "tn", trans_c=True),
+        ref_fn=_ref,
+        acc_slice=slice(None),
+    )
+    return runner.run(spec)
+
+
+def _dispatch_profile(group, args):
+    ctx = _dispatch_make_context(group, args)
+    runner = _make_runner(group, args)
+    checks = {}
+    try:
+        fwd_metrics, checks["fwd"], xgmi_bytes = _dispatch_stage_fwd(runner, ctx)
+        bwd_metrics, checks["bwd"] = _dispatch_stage_bwd_dgrad(runner, ctx)
+        wgrad_metrics, checks["wgrad"] = _dispatch_stage_bwd_wgrad(runner, ctx)
+    finally:
+        ctx.symm.destroy()  # always free symmetric buffers
+    # raw per-rank timings + work; rank 0 aggregates across ranks (bottleneck = max latency)
+    return {
+        "stages": {"fwd": fwd_metrics, "bwd": bwd_metrics, "wgrad": wgrad_metrics},
+        "xgmi_bytes": xgmi_bytes,
+        "checks": checks,
+    }
+
+
+###############################################################################
+# Mode: grouped_gemm_combine
+###############################################################################
+
+
+@dataclass
+class CombineContext:
+    """Shared per-run state built by _combine_make_context: config, input/symm buffers, derived geometry + tables."""
+
+    # collective + CLI config
+    group: Any
+    args: Any
+    rank: int
+    # input namespace + symmetric buffers (own act / handle / l2 buffers, read off inp/symm)
+    inp: Any
+    symm: Any
+    # derived geometry + tables (computed once, shared by both stages)
+    M_eff: int  # total padded pool rows this rank GEMMs
+    xgmi_bytes: int  # combine push bytes per rank (same fwd/bwd, H-wide)
+    num_tokens: int
+    topk_idx_flat: Any  # int32 [T*K], drives the per-token reduce
+    topk_w_flat: Any  # f32 [T*K], forward routing weights
+    reduce_ready: Any  # standalone reduce ready flags (0 == ready)
+
+
+def _combine_make_context(group, args):
+    """Build input + symm buffers and precompute shared geometry + topk tables (weights built in generate_input)."""
+    inp = generate_input(
+        group,
+        kind="combine",
+        T=args.num_tokens,
+        H=args.hidden,
+        I=args.inter,
+        E=args.num_experts,
+        K=args.num_topk,
+        BLOCK_M=256,
+        BLOCK_N=256,
+    )
+    symm = inp.symm
+    rank = group.rank()
+    real_tiles = int(inp.num_tile_blocks.item())
+    M_eff = real_tiles * 256
+    # combine push bytes per rank = remote rows (origin_rank != rank, valid) x H x bf16
+    origin = symm.pool_src_rank
+    remote_rows = int(((origin != rank) & (origin >= 0)).sum().item())
+    xgmi_bytes = remote_rows * args.hidden * BF16_BYTES
+    topk_idx_flat = inp.topk_idx.to(torch.int32).contiguous().view(-1)
+    topk_w_flat = inp.topk_weight.to(torch.float32).contiguous().view(-1)
+    reduce_ready = torch.zeros(int(symm.num_combine_slots), dtype=torch.int32, device="cuda")
+    return CombineContext(
+        group=group,
+        args=args,
+        rank=rank,
+        inp=inp,
+        symm=symm,
+        M_eff=M_eff,
+        xgmi_bytes=xgmi_bytes,
+        num_tokens=int(symm.num_tokens),
+        topk_idx_flat=topk_idx_flat,
+        topk_w_flat=topk_w_flat,
+        reduce_ready=reduce_ready,
+    )
+
+
+def _combine_make_fused_call(ctx, lhs, rhs, *, layout="nt", topk_weights):
+    """Build the fused GEMM + combine PUSH + topk-reduce call (3-role); stages differ in operands / layout / weights."""
+    return lambda: grouped_gemm_combine_bf16_flydsl_kernel(
+        lhs,
+        rhs,
+        ctx.inp.handle,
+        topk_indices=ctx.topk_idx_flat,
+        topk_weights=topk_weights,
+        layout=layout,
+        BM=256,
+        BN=256,
+    )
+
+
+def _combine_make_comm_call(ctx):
+    """Build the combine-only call (the comm baseline; CU count autotuned)."""
+    return lambda: combine_only(ctx.group, handle=ctx.inp.handle)
+
+
+def _combine_stage_fwd(runner, ctx):
+    """forward (e2e step 4, NT): L2 GEMM N=H, K=I; 3-role fused -> weighted y[T,H]."""
+    inp, args, symm = ctx.inp, ctx.args, ctx.symm
+    N, K = args.hidden, args.inter
+    flops = 2.0 * ctx.M_eff * N * K
+    ref_y = torch.empty(ctx.num_tokens, args.hidden, device="cuda", dtype=torch.bfloat16)
+
+    # reference: decoupled gemm_only(nt) + combine_only + weighted topk_reduce, over the
+    # SAME handle/buffers as fused_fn -> never-reset scoreboard drift cancels between them.
+    def _ref_fwd():
+        grouped_gemm_bf16_only(
+            inp.act,
+            inp.W2,
+            symm.l2_token_buffer,
+            inp.tile_to_expert,
+            inp.num_tile_blocks,
+            BLOCK_M=256,
+            BLOCK_N=256,
+            GROUP_M=8,
+        )
+        sync_ranks(ctx.group)
+        combine_only(ctx.group, handle=ctx.inp.handle)
+        sync_ranks(ctx.group)
+        ctx.reduce_ready.zero_()  # 0 == ready (reduce_only does not spin)
+        topk_reduce_only(
+            ref_y,
+            symm.combine_token_buffer,
+            ctx.reduce_ready,
+            ctx.topk_idx_flat,
+            symm.num_tokens_per_rank,
+            int(symm.num_combine_slots),
+            topk=int(symm.num_topk),
+            num_experts=int(symm.num_experts),
+            rank=ctx.rank,
+            topk_weights=ctx.topk_w_flat,  # weighted (forward routing weights)
+        )
+        return ref_y
+
+    # L2 has small K=I -> GROUP_M=8 reuses the per-expert weight across more M-tiles (best here)
+    spec = StageSpec(
+        name="fwd fused (nt)",
+        flops=flops,
+        dense_dims=(ctx.M_eff, N, K),
+        gemm_fn=lambda: grouped_gemm_bf16_only(
+            inp.act,
+            inp.W2,
+            symm.l2_token_buffer,
+            inp.tile_to_expert,
+            inp.num_tile_blocks,
+            BLOCK_M=256,
+            BLOCK_N=256,
+            GROUP_M=8,
+        ),
+        comm_fn=_combine_make_comm_call(ctx),
+        fused_fn=_combine_make_fused_call(ctx, inp.act, inp.W2, topk_weights=ctx.topk_w_flat),
+        ref_fn=_ref_fwd,
+    )
+    return runner.run(spec)
+
+
+def _combine_stage_bwd(runner, ctx):
+    """backward (e2e step 3, NN): L1 dgrad grad_l1[M,2I] @ w1 -> grad_pool[M,H]; 3-role fused -> dx[T,H]."""
+    inp, args, symm = ctx.inp, ctx.args, ctx.symm
+    N, K = args.hidden, 2 * args.inter  # weight [G, K=2I, N=H]
+    flops = 2.0 * ctx.M_eff * N * K
+    grad_l1 = torch.randn(symm.num_max_pool_tokens, K, device="cuda", dtype=torch.bfloat16) / 8
+    ref_dx = torch.empty(ctx.num_tokens, args.hidden, device="cuda", dtype=torch.bfloat16)
+
+    # reference: decoupled gemm_only(nn) + combine_only + unweighted topk_reduce (weight rides grad_l1)
+    def _ref_bwd():
+        grouped_gemm_bf16_only(
+            grad_l1,
+            inp.W1,
+            symm.l2_token_buffer,
+            inp.tile_to_expert,
+            inp.num_tile_blocks,
+            layout="nn",
+            BLOCK_M=256,
+            BLOCK_N=256,
+            GROUP_M=8,
+        )
+        sync_ranks(ctx.group)
+        combine_only(ctx.group, handle=ctx.inp.handle)
+        sync_ranks(ctx.group)
+        ctx.reduce_ready.zero_()  # 0 == ready (reduce_only does not spin)
+        topk_reduce_only(
+            ref_dx,
+            symm.combine_token_buffer,
+            ctx.reduce_ready,
+            ctx.topk_idx_flat,
+            symm.num_tokens_per_rank,
+            int(symm.num_combine_slots),
+            topk=int(symm.num_topk),
+            num_experts=int(symm.num_experts),
+            rank=ctx.rank,
+            topk_weights=None,  # unweighted (weight rides grad_l1)
+        )
+        return ref_dx
+
+    spec = StageSpec(
+        name="bwd dgrad fused (nn)",
+        flops=flops,
+        dense_dims=(ctx.M_eff, N, K),
+        gemm_fn=lambda: grouped_gemm_bf16_only(
+            grad_l1,
+            inp.W1,
+            symm.l2_token_buffer,
+            inp.tile_to_expert,
+            inp.num_tile_blocks,
+            layout="nn",
+            BLOCK_M=256,
+            BLOCK_N=256,
+            GROUP_M=8,
+        ),
+        comm_fn=_combine_make_comm_call(ctx),
+        fused_fn=_combine_make_fused_call(ctx, grad_l1, inp.W1, layout="nn", topk_weights=None),
+        ref_fn=_ref_bwd,
+    )
+    return runner.run(spec)
+
+
+def _combine_profile(group, args):
+    """Forward = e2e step 4 (NT weighted); backward = e2e step 3 (NN unweighted). Both report the 3-role fused kernel."""
+    ctx = _combine_make_context(group, args)
+    runner = _make_runner(group, args)
+    checks = {}
+    try:
+        fwd_metrics, checks["fwd"] = _combine_stage_fwd(runner, ctx)
+        bwd_metrics, checks["bwd"] = _combine_stage_bwd(runner, ctx)
+        xgmi_bytes = ctx.xgmi_bytes
+    finally:
+        ctx.symm.destroy()  # always free the symmetric buffer
+    # raw per-rank timings + work; rank 0 aggregates across ranks (bottleneck = max latency)
+    return {
+        "stages": {"fwd": fwd_metrics, "bwd": bwd_metrics},
+        "xgmi_bytes": xgmi_bytes,
+        "checks": checks,
+    }
+
+
+###############################################################################
+# Mode registry + shared reporting / driver
+###############################################################################
+
+
+@dataclass
+class StageReport:
+    """Per-stage reporting descriptor: which result key, its column prefix / flags, print sub_header."""
+
+    key: str  # per_rank["stages"] / checks key
+    col_prefix: str  # stage_columns name prefix ("", "bwd ", "wgrad ")
+    col_flags: dict = field(default_factory=dict)  # extra stage_columns switches (dense_ms / xgmi / hidden)
+    sub_header: str = ""  # print_stage sub_header ("" = none, used by the fwd stage)
+
+
+@dataclass
+class ModeSpec:
+    """Everything the shared driver needs to run + report one mode."""
+
+    profile: Callable  # (group, args) -> {"stages", "xgmi_bytes", "checks"}
+    port: int  # default MASTER_PORT for this mode's spawn
+    prefix: str  # output CSV filename prefix
+    header_kind: str  # print_header kind label
+    comm_label: str  # comm baseline name (dispatch_only / combine_only)
+    comm_unit: str  # comm bandwidth unit string
+    comm_tag: str  # short comm tag (disp / comb); reused as stage_columns comm_short
+    stages: list  # ordered StageReport; [0] is the fwd (csv fwd), the rest are bwd_stages
+
+
+# fwd / bwd share the same column flags across modes; only labels + the extra wgrad stage differ
+_FWD_REPORT = StageReport("fwd", "", {"dense_ms": True, "xgmi": True, "hidden": True})
+
+MODES = {
+    "dispatch_grouped_gemm": ModeSpec(
+        profile=_dispatch_profile,
+        port=8481,
+        prefix="dispatch_grouped_gemm",
+        header_kind="dispatch",
+        comm_label="dispatch_only",
+        comm_unit="GB/s (XGMI, nodeup)",
+        comm_tag="disp",
+        stages=[
+            _FWD_REPORT,
+            StageReport("bwd", "bwd ", {"xgmi": True}, "backward dgrad (NN, = dispatch_grouped_0)"),
+            StageReport(
+                "wgrad", "wgrad ", {}, "backward wgrad dW1 (TN, = dispatch + variable-K wgrad)"
+            ),
+        ],
+    ),
+    "grouped_gemm_combine": ModeSpec(
+        profile=_combine_profile,
+        port=8483,
+        prefix="grouped_gemm_combine",
+        header_kind="combine",
+        comm_label="combine_only",
+        comm_unit="GB/s (XGMI)",
+        comm_tag="comb",
+        stages=[
+            _FWD_REPORT,
+            StageReport("bwd", "bwd ", {"xgmi": True}, "backward dgrad (NN, = mega_moe_fused STEP 3)"),
+        ],
+    ),
+}
+
+
+def _report_case(mode, platform, gpu_name, world, args, case_name, per_rank):
+    """rank-0: aggregate one case's per-rank results -> (rich_row, csv_row); prints the detail block."""
+    # distributed bottleneck = slowest rank (max latency); work ~ uniform (mean)
+    xgmi = statistics.mean([res["xgmi_bytes"] for res in per_rank])
+    rank0_checks = per_rank[0]["checks"]
+    aggs = {s.key: aggregate_stage_metrics(per_rank, s.key, xgmi) for s in mode.stages}
+
+    print_header(mode.header_kind, gpu_name, world, args, case=case_name)
+    for s in mode.stages:
+        extra = {"sub_header": s.sub_header} if s.sub_header else {}
+        print_stage(
+            aggs[s.key],
+            comm_label=mode.comm_label,
+            comm_unit=mode.comm_unit,
+            comm_tag=mode.comm_tag,
+            **extra,
+        )
+    rich_row = {
+        "Platform": platform,
+        "GPU": gpu_name,
+        "Case": case_name,
+        "EP": world,
+        "T": args.num_tokens,
+        "H": args.hidden,
+        "I": args.inter,
+        "E": args.num_experts,
+        "K": args.num_topk,
+    }
+    for s in mode.stages:
+        rich_row.update(
+            stage_columns(
+                s.col_prefix,
+                aggs[s.key],
+                rank0_checks.get(s.key),
+                comm_label=mode.comm_label,
+                comm_short=mode.comm_tag,
+                **s.col_flags,
+            )
+        )
+    # saved CSV follows the gemm_turbo convention; fwd = stages[0], backward = the rest
+    csv_row = turbo_csv_row(
+        platform,
+        gpu_name,
+        world,
+        args,
+        case=case_name,
+        check=checks_verdict(rank0_checks),
+        fwd=aggs[mode.stages[0].key],
+        bwd_stages=[aggs[s.key] for s in mode.stages[1:]],
+    )
+    return rich_row, csv_row
+
+
+def _init_dist(local_rank, world, default_port):
+    """Bring up NCCL + the collective group for this spawn; returns the process group."""
+    master_addr = os.getenv("MASTER_ADDR", "127.0.0.1")
+    port = int(os.getenv("MASTER_PORT", str(default_port)))
+    torch.cuda.set_device(local_rank)
+    dist.init_process_group(
+        "nccl", init_method=f"tcp://{master_addr}:{port}", world_size=world, rank=local_rank
+    )
+    torch.set_default_device("cuda")
+    torch.cuda.set_device(local_rank)
+    return dist.new_group(list(range(world)))
+
+
+def _benchmark(local_rank, world, args):
+    """One spawn sweeps every MoE model for args.mode; unsupported / failing cases are skipped."""
+    mode = MODES[args.mode]
+    group = _init_dist(local_rank, world, mode.port)
+    rank = dist.get_rank()
+    platform, gpu_name = get_platform_info()
+
+    try:
+        rich_rows, csv_rows = [], []
+        for case in gen_moe_test_cases(args.models):
+            name = case["Case"]
+            if name in UNSUPPORTED:
+                if rank == 0:
+                    print(f"[skip] {name}: unsupported by mega {args.mode} (see UNSUPPORTED TODO)")
+                continue
+            apply_case(args, case)
+            sync_ranks(group)
+            ok, result = True, None
+            try:
+                result = mode.profile(group, args)
+            except Exception as e:  # noqa: BLE001  probe: skip cases the kernel can't run
+                ok = False
+                if rank == 0:
+                    print(f"[skip] {name}: {e!r}")
+            # collective agreement so every rank skips a failed case together (no hang)
+            if not all_ranks_ok(group, ok):
+                sync_ranks(group)
+                continue
+            per_rank = [None] * world  # all_gather_object fills per_rank[i] from rank i
+            dist.all_gather_object(per_rank, result, group=group)
+            if rank == 0:
+                rich_row, csv_row = _report_case(mode, platform, gpu_name, world, args, name, per_rank)
+                rich_rows.append(rich_row)
+                csv_rows.append(csv_row)
+            sync_ranks(group)
+
+        if rank == 0 and csv_rows:
+            print("\nFinal Results:")
+            print(tabulate(pd.DataFrame(rich_rows), headers="keys", tablefmt="grid", showindex=False))
+            out_file = args.output or f"{mode.prefix}_{datetime.now():%Y%m%d}_{gpu_name}.csv"
+            pd.DataFrame(csv_rows).to_csv(out_file, index=False)
+            print(f"Results saved to {out_file}")
+        sync_ranks(group)
+    finally:
+        dist.destroy_process_group()
+
+
+def _run(args):
+    """Driver: one spawn iterates all cases in-process for the selected mode."""
+    torch.multiprocessing.spawn(_benchmark, args=(args.num_processes, args), nprocs=args.num_processes)
+
+
+###############################################################################
+# Entrypoint
+###############################################################################
+
+
+def _build_parser():
+    parser = argparse.ArgumentParser(description="Benchmark the fused BF16 mega MoE kernels")
+    parser.add_argument(
+        "--mode",
+        choices=list(MODES),
+        required=True,
+        help="which fused mega kernel to benchmark",
+    )
+    parser.add_argument("--num-processes", type=int, default=8)
+    # H/I/E/K come from each MoE model case (config.gen_moe_test_cases); no CLI knob
+    parser.add_argument("--num-tokens", type=int, default=8192)
+    # GROUP_M for the dense roofline reference
+    parser.add_argument("--dense-group-m", type=int, default=4)
+    parser.add_argument("--iters", type=int, default=30)
+    parser.add_argument("--output", "-o", type=str, default=None)
+    # restrict the sweep to these MoE model names (default = all)
+    parser.add_argument("--models", nargs="+", default=None)
+    return parser
+
+
+if __name__ == "__main__":
+    args = _build_parser().parse_args()
+    _run(args)
